- restructured private methods
