+++ Browser/Classes/WebView2Service/Class_WV2Service.Private.History.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Diagnostics;
+using Microsoft.Web.WebView2.Core;
+using Microsoft.Web.WebView2.WinForms;
+
+namespace WV2Service
+{
+    public partial class WebViewService
+    {
+    }
+}
