+++ Browser/Classes/WebView2Service/Class_WV2Service.Private.cs
@@ -1,12 +1,6 @@
 ﻿using Microsoft.Web.WebView2.Core;
 using Microsoft.Web.WebView2.WinForms;
-using System.ComponentModel;
-using System.Data;
 using System.Diagnostics;
-using System.Reflection;
-using System.Runtime.CompilerServices;
-using System.Threading;
-using System.Windows.Forms;
 
 namespace WV2Service
 {
@@ -89,6 +83,17 @@ namespace WV2Service
                 throw new Exception($"Failed to initialize environment:\n{ex.Message}");
             }
         }
+        private async void EnableMobileView(WebView2 webView, CoreWebView2Environment environment)
+        {
+            await webView.EnsureCoreWebView2Async(environment);
+            webView.CoreWebView2.Settings.UserAgent = @"Mozilla/5.0 (Linux; Android 10; Mobile) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/108.0.0.0 Mobile Safari/537.36";
+        }
+        private async void InitializeExtensions(WebView2 webView, CoreWebView2Environment environment, CoreWebView2Profile profile)
+        {
+            await webView.EnsureCoreWebView2Async(environment);
+            CoreWebView2BrowserExtension extension = await AddExtensionsAsync(webView, environment, profile);
+            await extension.EnableAsync(true);
+        }
         private async Task<CoreWebView2BrowserExtension> AddExtensionsAsync(WebView2 webView, CoreWebView2Environment environment, CoreWebView2Profile profile)
         {
             try
@@ -124,155 +129,6 @@ namespace WV2Service
             {
                 throw new Exception($"Error retrieving extensions:\n{ex.Message}");
             }
-        }
-        private async void EnableMobileView(WebView2 webView, CoreWebView2Environment environment)
-        {
-            await webView.EnsureCoreWebView2Async(environment);
-            webView.CoreWebView2.Settings.UserAgent = @"Mozilla/5.0 (Linux; Android 10; Mobile) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/108.0.0.0 Mobile Safari/537.36";
-        }
-        private async void InitializeExtensions(WebView2 webView, CoreWebView2Environment environment, CoreWebView2Profile profile)
-        {
-            await webView.EnsureCoreWebView2Async(environment);
-            CoreWebView2BrowserExtension extension = await AddExtensionsAsync(webView, environment, profile);
-            await extension.EnableAsync(true);
-        }
-        private async void ClearAllBrowsingData(WebView2 webView, CoreWebView2Environment environment, CoreWebView2Profile profile)
-        {
-            try
-            {
-                profile = profile != null? profile : await GetProfile(webView, environment);
-                await profile.ClearBrowsingDataAsync(
-                    CoreWebView2BrowsingDataKinds.Cookies |
-                    CoreWebView2BrowsingDataKinds.BrowsingHistory |
-                    CoreWebView2BrowsingDataKinds.GeneralAutofill |
-                    CoreWebView2BrowsingDataKinds.PasswordAutosave |
-                    CoreWebView2BrowsingDataKinds.ServiceWorkers |
-                    CoreWebView2BrowsingDataKinds.CacheStorage |
-                    CoreWebView2BrowsingDataKinds.DownloadHistory |
-                    CoreWebView2BrowsingDataKinds.DiskCache);
-                webView.Reload();
-            }
-            catch (Exception ex)
-            {
-                throw new Exception($"Failed to clear data:\n{ex.Message}");
-            }
-        }
-        private async void ClearBrowsingDataBetweenDateRange(WebView2 webView, CoreWebView2Environment environment, CoreWebView2Profile profile, DateTime startDate, DateTime endDate)
-        {
-            try
-            {
-                profile = profile != null ? profile : await GetProfile(webView, environment);
-                await profile.ClearBrowsingDataAsync(
-                    CoreWebView2BrowsingDataKinds.Cookies |
-                    CoreWebView2BrowsingDataKinds.BrowsingHistory |
-                    CoreWebView2BrowsingDataKinds.GeneralAutofill |
-                    CoreWebView2BrowsingDataKinds.PasswordAutosave |
-                    CoreWebView2BrowsingDataKinds.ServiceWorkers |
-                    CoreWebView2BrowsingDataKinds.CacheStorage |
-                    CoreWebView2BrowsingDataKinds.DownloadHistory |
-                    CoreWebView2BrowsingDataKinds.DiskCache, startDate, endDate);
-                webView.Reload();
-            }
-            catch (Exception ex)
-            {
-                throw new Exception($"Failed to clear data:\n{ex.Message}");
-            }
-        }
-        private async void ClearBrowserData(WebView2 webView, CoreWebView2Environment environment, CoreWebView2Profile profile, CoreWebView2BrowsingDataKinds dataKinds)
-        {
-            try
-            {
-                profile = profile != null ? profile : await GetProfile(webView, environment);
-                await profile.ClearBrowsingDataAsync(dataKinds);
-                webView.Reload();
-            }
-            catch (Exception ex)
-            {
-                throw new Exception($"Failed to clear data:\n{ex.Message}");
-            }
-        }
-        private async void ClearAllBrowserData(WebView2 webView, CoreWebView2Environment environment, CoreWebView2Profile profile)
-        {
-            profile = profile != null ? profile : await GetProfile(webView, environment);
-            await webView.EnsureCoreWebView2Async(environment);
-            await profile.ClearBrowsingDataAsync();
-            webView.Reload();
-        }
-        private string EnsureHttpsPrefix(string url)
-        {
-            if (!url.StartsWith("http://", StringComparison.OrdinalIgnoreCase) &&
-                !url.StartsWith("https://", StringComparison.OrdinalIgnoreCase) &&
-                !url.StartsWith("edge://", StringComparison.OrdinalIgnoreCase))
-            {
-                url = "https://" + url;
-            }
-            return url;
-        }
-        private bool IsURLSuffixValid(string url)
-        {
-            string[] validTLDs = { ".com", ".org", ".net", ".edu", ".gov", ".io", ".co", ".us", ".uk", ".ph" };
-            if (url.StartsWith("edge://", StringComparison.OrdinalIgnoreCase))  { return true; }
-            foreach (string tld in validTLDs)
-            {
-                if (url.Contains(tld, StringComparison.OrdinalIgnoreCase))
-                {
-                    return true;
-                }
-            }
-            return false;
-        }
-        private async void NavigateTo(WebView2 webView, CoreWebView2Environment environment, string address)
-        {
-            await webView.EnsureCoreWebView2Async(environment);
-            webView.CoreWebView2.Navigate(address);
-        }
-        private async void EnableNewWindowRequest(WebView2 webView, CoreWebView2Environment environment)
-        {
-            await webView.EnsureCoreWebView2Async(environment);
-            webView.CoreWebView2.NewWindowRequested += CoreWebView2_NewWindowRequested;
-        }
-        private async void EnableNavigationMonitoring(WebView2 webView, CoreWebView2Environment environment)
-        {
-            await webView.EnsureCoreWebView2Async(environment);
-            webView.NavigationStarting += OnNavigationStarting;
-            webView.NavigationCompleted += OnNavigationCompleted;
-        }
-        private void CoreWebView2_NewWindowRequested(object? sender, CoreWebView2NewWindowRequestedEventArgs e)
-        {
-            NewWindowRequested?.Invoke(sender, e);
-
-            if (NewWindowRequested == null)
-            {
-                DialogResult result = MessageBox.Show( $"A website wants to open a new window:\n{e.Uri}\n\nAllow this popup?", "Confirm Popup",  MessageBoxButtons.YesNo, MessageBoxIcon.Question );
-
-                if (result == DialogResult.Yes)
-                {
-                    e.Handled = false;
-                    //// Optional: Open in the current WebView
-                    //NavigateTo(webviewControl, environment, e.Uri);
-                    return;
-                }
-                e.Handled = true;
-                Debug.WriteLine($"Popup blocked: {e.Uri}");
-            }
-        }
-        private async void OnNavigationCompleted(object? sender, CoreWebView2NavigationCompletedEventArgs e)
-        {
-            if (e.IsSuccess)
-            {
-                string siteTitle = await WebViewControl.CoreWebView2.ExecuteScriptAsync("document.title");
-                SiteTitle = siteTitle.Trim('"');
-                NavigationChanged?.Invoke(this, WebViewControl.Source.ToString());
-            }
-            else
-            {
-                NavigationChanged?.Invoke(this, "Navigation failed.");
-            }
-        } 
-        private void OnNavigationStarting(object? sender, CoreWebView2NavigationStartingEventArgs e)
-        {
-            URL = e.Uri;
-            NavigationChanged?.Invoke(this, $"Navigating to: {e.Uri}");
         }
     }
 }
