+++ Browser/Classes/WebView2Service/Class_WV2Service.Public.Navigation.cs
@@ -13,37 +13,11 @@ namespace WV2Service
             private readonly WebViewService wv2service = webviewService;
             public void GoTo(string address)
             {
-                if (wv2service.IsURLSuffixValid(address))
-                {
-                    wv2service.URL = wv2service.EnsureHttpsPrefix(address);
-                    wv2service.NavigateTo(wv2service.WebViewControl, wv2service.environment, wv2service.URL);
-                    return;
-                }
-
-                string searchQuery = Uri.EscapeDataString(address);
-                string searchUrl = "https://www.google.com/search?q=" + searchQuery;
-
-                wv2service.URL = (new Uri(searchUrl)).ToString();
-                wv2service.NavigateTo(wv2service.WebViewControl, wv2service.environment, wv2service.URL);
+                wv2service.NavigateTo(wv2service.WebViewControl, wv2service.environment, address);
             }
             public void NewTabGoTo(string address)
             {
-                if (wv2service.IsURLSuffixValid(address))
-                {
-                    //address = EnsureHttpsPrefix(address);
-                    //WebViewControl.EnsureCoreWebView2Async();
-                    //WebViewControl.Source = new Uri(address);
-
-                    wv2service.URL = wv2service.EnsureHttpsPrefix(address);
-                    wv2service.NavigateTo(wv2service.WebViewControl, wv2service.environment, wv2service.URL);
-                    return;
-                }
-
-                string searchQuery = Uri.EscapeDataString(address);
-                string searchUrl = "https://www.google.com/search?q=" + searchQuery;
-
-                wv2service.WebViewControl.EnsureCoreWebView2Async();
-                wv2service.WebViewControl.Source = new Uri(searchUrl);
+                wv2service.NavigateToNewTab(wv2service.WebViewControl, wv2service.environment, address);
             }
             public void Reload()
             {
